requirements.txt
python-decouple
Telethon
argparse